# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "goose-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.19.0"
  },
  "paths": {
...
    "/notes/create": {
      "post": {
        "tags": [
          "Notes"
        ],
        "operationId": "create_note",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Note created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateNoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/notes/list": {
      "get": {
        "tags": [
          "Notes"
        ],
        "operationId": "list_notes",
        "responses": {
          "200": {
            "description": "Notes retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListNotesResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/notes/{id}": {
      "get": {
        "tags": [
          "Notes"
        ],
        "operationId": "get_note",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Note ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetNoteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Note not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Notes"
        ],
        "operationId": "update_note",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Note ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Note updated successfully"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Note not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Notes"
        ],
        "operationId": "delete_note",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Note ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note deleted successfully"
          },
          "404": {
            "description": "Note not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/sessions/{session_id}": {
      "get": {
        "tags": [
          "Session Management"
        ],
        "operationId": "get_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Session Management"
        ],
        "operationId": "delete_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session deleted successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Author": {
        "type": "object",
        "properties": {
          "contact": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Conversation": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Message"
        }
      },
      "CreateNoteRequest": {
        "type": "object",
        "required": [
          "session_ids"
        ],
        "properties": {
          "session_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateNoteResponse": {
        "type": "object",
        "required": [
          "note"
        ],
        "properties": {
          "note": {
            "$ref": "#/components/schemas/Note"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "ExtensionData": {
        "type": "object",
        "description": "Extension data containing all extension states\nKeys are in format \"extension_name.version\" (e.g., \"todo.v0\")",
        "additionalProperties": {}
      },
...
      "GetNoteResponse": {
        "type": "object",
        "required": [
          "id",
          "title",
          "content",
          "citations"
        ],
        "properties": {
          "citations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoteCitationResponse"
            }
          },
          "content": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "ListNotesResponse": {
        "type": "object",
        "required": [
          "notes"
        ],
        "properties": {
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoteListItem"
            }
          }
        }
      },
...
      "Message": {
        "type": "object",
        "description": "A message to or from an LLM",
        "required": [
          "role",
          "created",
          "content",
          "metadata"
        ],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageContent"
            }
          },
          "created": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "$ref": "#/components/schemas/MessageMetadata"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
...
      "MessageMetadata": {
        "type": "object",
        "description": "Metadata for message visibility",
        "required": [
          "userVisible",
          "agentVisible"
        ],
        "properties": {
          "agentVisible": {
            "type": "boolean",
            "description": "Whether the message should be included in the agent's context window"
          },
          "userVisible": {
            "type": "boolean",
            "description": "Whether the message should be visible to the user in the UI"
          }
        }
      },
      "ModelConfig": {
        "type": "object",
        "required": [
          "model_name",
          "toolshim"
        ],
        "properties": {
          "context_limit": {
            "type": "integer",
            "nullable": true,
            "minimum": 0
          },
          "fast_model": {
            "type": "string",
            "nullable": true
          },
          "max_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "model_name": {
            "type": "string"
          },
          "temperature": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "toolshim": {
            "type": "boolean"
          },
          "toolshim_model": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Note": {
        "type": "object",
        "required": [
          "id",
          "title",
          "content",
          "citations"
        ],
        "properties": {
          "citations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoteCitation"
            }
          },
          "content": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "NoteCitation": {
        "type": "object",
        "required": [
          "session_id",
          "message_id",
          "citation_index"
        ],
        "properties": {
          "citation_index": {
            "type": "integer",
            "format": "int32"
          },
          "message_id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          }
        }
      },
      "NoteCitationResponse": {
        "type": "object",
        "required": [
          "session_id",
          "message_id",
          "citation_index"
        ],
        "properties": {
          "citation_index": {
            "type": "integer",
            "format": "int32"
          },
          "message_id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          }
        }
      },
      "NoteListItem": {
        "type": "object",
        "required": [
          "id",
          "title",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
      "Recipe": {
        "type": "object",
        "required": [
          "title",
          "description"
        ],
        "properties": {
          "activities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "author": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Author"
              }
            ],
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "extensions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExtensionConfig"
            },
            "nullable": true
          },
          "instructions": {
            "type": "string",
            "nullable": true
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeParameter"
            },
            "nullable": true
          },
          "prompt": {
            "type": "string",
            "nullable": true
          },
          "response": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Response"
              }
            ],
            "nullable": true
          },
          "retry": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RetryConfig"
              }
            ],
            "nullable": true
          },
          "settings": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Settings"
              }
            ],
            "nullable": true
          },
          "sub_recipes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubRecipe"
            },
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "RecipeParameter": {
        "type": "object",
        "required": [
          "key",
          "input_type",
          "requirement",
          "description"
        ],
        "properties": {
          "default": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "input_type": {
            "$ref": "#/components/schemas/RecipeParameterInputType"
          },
          "key": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "requirement": {
            "$ref": "#/components/schemas/RecipeParameterRequirement"
          }
        }
      },
      "RecipeParameterInputType": {
        "type": "string",
        "enum": [
          "string",
          "number",
          "boolean",
          "date",
          "file",
          "select"
        ]
      },
      "RecipeParameterRequirement": {
        "type": "string",
        "enum": [
          "required",
          "optional",
          "user_prompt"
        ]
      },
...
      "Response": {
        "type": "object",
        "properties": {
          "json_schema": {
            "nullable": true
          }
        }
      },
...
      "RetryConfig": {
        "type": "object",
        "description": "Configuration for retry logic in recipe execution",
        "required": [
          "max_retries",
          "checks"
        ],
        "properties": {
          "checks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SuccessCheck"
            },
            "description": "List of success checks to validate recipe completion"
          },
          "max_retries": {
            "type": "integer",
            "format": "int32",
            "description": "Maximum number of retry attempts before giving up",
            "minimum": 0
          },
          "on_failure": {
            "type": "string",
            "description": "Optional shell command to run on failure for cleanup",
            "nullable": true
          },
          "on_failure_timeout_seconds": {
            "type": "integer",
            "format": "int64",
            "description": "Timeout in seconds for on_failure commands (default: 600 seconds)",
            "nullable": true,
            "minimum": 0
          },
          "timeout_seconds": {
            "type": "integer",
            "format": "int64",
            "description": "Timeout in seconds for individual shell commands (default: 300 seconds)",
            "nullable": true,
            "minimum": 0
          }
        }
      },
      "Role": {
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "string"
          }
        ]
      },
...
      "Session": {
        "type": "object",
        "required": [
          "id",
          "working_dir",
          "name",
          "created_at",
          "updated_at",
          "extension_data",
          "message_count"
        ],
        "properties": {
          "accumulated_input_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "accumulated_output_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "accumulated_total_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "conversation": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Conversation"
              }
            ],
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "extension_data": {
            "$ref": "#/components/schemas/ExtensionData"
          },
          "id": {
            "type": "string"
          },
          "input_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "message_count": {
            "type": "integer",
            "minimum": 0
          },
          "model_config": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ModelConfig"
              }
            ],
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "output_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "provider_name": {
            "type": "string",
            "nullable": true
          },
          "recipe": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Recipe"
              }
            ],
            "nullable": true
          },
          "schedule_id": {
            "type": "string",
            "nullable": true
          },
          "session_type": {
            "$ref": "#/components/schemas/SessionType"
          },
          "total_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_recipe_values": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "user_set_name": {
            "type": "boolean"
          },
          "working_dir": {
            "type": "string"
          }
        }
      },
...
      "SessionType": {
        "type": "string",
        "enum": [
          "user",
          "scheduled",
          "sub_agent",
          "hidden",
          "terminal"
        ]
      },
...
      "Settings": {
        "type": "object",
        "properties": {
          "goose_model": {
            "type": "string",
            "nullable": true
          },
          "goose_provider": {
            "type": "string",
            "nullable": true
          },
          "temperature": {
            "type": "number",
            "format": "float",
            "nullable": true
          }
        }
      },
...
      "SubRecipe": {
        "type": "object",
        "required": [
          "name",
          "path"
        ],
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "sequential_when_repeated": {
            "type": "boolean"
          },
          "values": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          }
        }
      },
      "SuccessCheck": {
        "oneOf": [
          {
            "type": "object",
            "description": "Execute a shell command and check its exit status",
            "required": [
              "command",
              "type"
            ],
            "properties": {
              "command": {
                "type": "string",
                "description": "The shell command to execute"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Shell"
                ]
              }
            }
          }
        ],
        "description": "A single success check to validate recipe completion",
        "discriminator": {
          "propertyName": "type"
        }
      },
...
      "UpdateNoteRequest": {
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
    }
  }
}